Refactor to move info into new files.
